refactor: clean up ExploreStory and SimilarStory components
