+++ src/components/home/NewestStories.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Bookmarks, CaretCircleRight } from '@phosphor-icons/react';
+import { CaretCircleRight } from '@phosphor-icons/react';
 import { useRouter } from 'next/navigation';
 import { useTranslations } from 'next-intl';
 import React from 'react';
@@ -23,43 +23,6 @@ const NewestStories = () => {
     sortBy: 'createdAt',
   });
 
-  const handleAddToFavorites = (storyId: number) => {
-    console.log('storyId', storyId);
-  };
-
-  const renderActions = (storyId: number) => {
-    return (
-      <div
-        className={mergeClassnames(
-          'flex w-full items-center gap-2 justify-self-end mt-3 absolute bottom-[10px]',
-          'md:flex-row md:mt-2 md:px-3 md:pl-0',
-        )}
-      >
-        <Button
-          variant="primary"
-          className={mergeClassnames(
-            'text-base h-8 max-h-8 w-[120px] flex-none rounded-full px-[12px] py-[12px]',
-            'md:h-[44px] md:max-h-[44px] md:w-[105px]',
-          )}
-          onClick={() => router.push(`/explore-story/${storyId}`)}
-        >
-          Read all
-        </Button>
-        <Button
-          variant="outline"
-          className={mergeClassnames(
-            'w-full h-8',
-            'md:size-10 md:min-h-10 md:min-w-10',
-          )}
-          iconOnly
-          onClick={() => handleAddToFavorites(storyId)}
-        >
-          <Bookmarks size={20} />
-        </Button>
-      </div>
-    );
-  };
-
   if (loadingStories) return <StoriesSkeleton />;
 
   return (
@@ -77,11 +40,7 @@ const NewestStories = () => {
         >
           {/* <div className="relative mt-4 flex flex-wrap items-center justify-center gap-8 xl:justify-start 2xl:gap-12"> */}
           {storiesPages?.data?.map((item: StoryType) => (
-            <FlipBook
-              key={item.id}
-              data={item}
-              renderActions={() => renderActions(item?.id)}
-            />
+            <FlipBook key={item.id} data={item} />
           ))}
         </div>
       ) : (
